+++ src/clear_completed.js
@@ -1,0 +1,7 @@
+import { listManager } from './list_manager';
+
+const clearCompleted = () => {
+  listManager.clearCompleted();
+};
+
+export { clearCompleted as default };
