+++ src/index.js
@@ -5,11 +5,8 @@ import '@fortawesome/fontawesome-free/js/regular';
 import '@fortawesome/fontawesome-free/js/brands';
 
 import ListListeners from './list_listeners';
-import DisplayList from './display_list';
-import ListManager from './list_manager';
 
-const listManager = new ListManager();
-const displayList = new DisplayList(listManager);
-displayList.render();
+import { listManager } from './list_manager';
+
 const listListeners = new ListListeners(listManager);
 listListeners.attachListeners();
